+++ test_all.sh
@@ -2,6 +2,8 @@
 
 set -x
 
+cargo build --no-default-features
+cargo build --no-default-features --feature "color"
 cargo test --all --verbose
 
 cd fuzz